OPENE-31 Fix for failing tests.
